+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/abap-on-btp/questions.test.ts
@@ -1,12 +1,13 @@
-import type { ServiceProvider } from '@sap-ux/axios-extension';
+import type { ServiceInfo, ServiceProvider } from '@sap-ux/axios-extension';
 import type { Question } from 'inquirer';
 import type { ListQuestion } from '@sap-ux/inquirer-common';
 import { apiGetServicesInstancesFilteredByType as getServicesFromCF, type ServiceInstanceInfo } from '@sap/cf-tools';
-import { ERROR_TYPE, ErrorHandler } from '../../../../../src/error-handler/error-handler';
+import { ERROR_TYPE, ErrorHandler } from '@sap-ux/inquirer-common';
 import { initI18nOdataServiceInquirer, t } from '../../../../../src/i18n';
 import type { ConnectionValidator } from '../../../../../src/prompts/connectionValidator';
 import { getAbapOnBTPSystemQuestions } from '../../../../../src/prompts/datasources/sap-system/abap-on-btp/questions';
 import { PromptState } from '../../../../../src/utils';
+import * as sapSystemValidators from '../../../../../src/prompts/datasources/sap-system/validators';
 
 const validateUrlMock = jest.fn().mockResolvedValue(true);
 const validateAuthMock = jest.fn().mockResolvedValue(true);
@@ -227,7 +228,10 @@ describe('questions', () => {
         const errorHandlerSpy = jest.spyOn(ErrorHandler.prototype, 'logErrorMsgs');
         PromptState.isYUI = true;
         expect(await ((cfDiscoPrompt as ListQuestion).choices as Function)()).toEqual([]);
-        expect(errorHandlerSpy).toHaveBeenCalledWith(ERROR_TYPE.NO_ABAP_ENVS, t('errors.noAbapEnvsInCFSpace'));
+        expect(errorHandlerSpy).toHaveBeenCalledWith(
+            ERROR_TYPE.NO_ABAP_ENVS,
+            'No ABAP environments in CF space found.'
+        );
         expect(await ((cfDiscoPrompt as ListQuestion).validate as Function)()).toEqual(
             'No ABAP environments in CF space found. See log for more details.'
         );
@@ -243,7 +247,8 @@ describe('questions', () => {
         (getServicesFromCF as jest.Mock).mockRejectedValueOnce(new Error('Not logged in'));
         expect(await ((cfDiscoPrompt as ListQuestion).choices as Function)()).toEqual([]);
         expect(await ((cfDiscoPrompt as ListQuestion).validate as Function)()).toEqual(
-            `${t('errors.abapEnvsCFDiscoveryFailed')} ${t('texts.seeLogForDetails')}`
+            'Discovering ABAP Environments failed. Please ensure you are logged into Cloud Foundry ' +
+                '(see https://docs.cloudfoundry.org/cf-cli/getting-started.html#login). See log for more details.'
         );
     });
 
@@ -327,5 +332,37 @@ describe('questions', () => {
         ).rejects.toThrowError('Cannot connect');
 
         expect(connectionValidatorMock.connectedSystemName).toBe(undefined);
+    });
+
+    test('Service key prompt should validate service key and connect', async () => {
+        const serviceInfoMock: ServiceInfo = {
+            uaa: {
+                clientid: 'clientid1',
+                clientsecret: 'clientSecret1',
+                url: 'url1'
+            },
+            url: 'url1',
+            catalogs: {
+                abap: {
+                    path: 'path1',
+                    type: 'type1'
+                }
+            }
+        };
+        let validateServiceKeyFileMock = jest
+            .spyOn(sapSystemValidators, 'validateServiceKey')
+            .mockReturnValue(serviceInfoMock); // service key file is valid
+        validateServiceInfoMock = true; // connection is successful
+        const newSystemQuestions = getAbapOnBTPSystemQuestions();
+
+        const serviceKeyPrompt = newSystemQuestions.find((q) => q.name === 'serviceKey');
+        expect(await (serviceKeyPrompt?.validate as Function)('path/to/service/key')).toBe(true);
+        expect(validateServiceKeyFileMock).toHaveBeenCalledWith('path/to/service/key');
+        expect(PromptState.odataService).toEqual({ connectedSystem: { serviceProvider: serviceProviderMock } });
+
+        validateServiceKeyFileMock = jest
+            .spyOn(sapSystemValidators, 'validateServiceKey')
+            .mockReturnValue('invalid service key file'); // service key file is valid
+        expect(await (serviceKeyPrompt?.validate as Function)('path/to/service/key')).toBe('invalid service key file');
     });
 });