+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/axios-extension
 
+## 1.17.2
+
+### Patch Changes
+
+-   Updated dependencies [3734fe8]
+    -   @sap-ux/btp-utils@0.16.0
+
+## 1.17.1
+
+### Patch Changes
+
+-   7551316: Small text cleanups
+
+## 1.17.0
+
+### Minor Changes
+
+-   0120dda: Handle HTTP(S) proxy configurations
+
 ## 1.16.7
 
 ### Patch Changes