+++ packages/control-property-editor-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux-private/control-property-editor-common
 
+## 0.5.7
+
+### Patch Changes
+
+-   8c0ba5c: Fixed Adaptation Editor crash when project contains Personalization change.
+
+## 0.5.6
+
+### Patch Changes
+
+-   06e9468: Allow adaptations of manifest settings in FEv4 adaptation projects via Control Property Editor Property Panel
+
 ## 0.5.5
 
 ### Patch Changes