+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.3.45
+
+### Patch Changes
+
+-   Updated dependencies [2886db3]
+    -   @sap-ux/inquirer-common@0.5.0
+
+## 0.3.44
+
+### Patch Changes
+
+-   Updated dependencies [fb26f92]
+    -   @sap-ux/project-access@1.28.6
+
+## 0.3.43
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.5
+
+## 0.3.42
+
+### Patch Changes
+
+-   Updated dependencies [5a68903]
+    -   @sap-ux/project-access@1.28.4
+
 ## 0.3.41
 
 ### Patch Changes